+++ modules/python_engine/tests/test_type_float.m
@@ -7,11 +7,8 @@
 % SPDX-License-Identifier: LGPL-3.0-or-later
 % LICENCE_BLOCK_END
 %=============================================================================
-function r = uint16(obj)
-  if ismethod(obj, 'uint16')
-    r = obj.uint16();
-  else
-    error([_('Wrong value for #2 argument.'), ' ', 'uint16']);
-  end
-end
+% <--PYTHON ENVIRONMENT REQUIRED-->
+%=============================================================================
+R = pyrun('R = 3.2', "R");
+assert_isequal(R, 3.2);
 %=============================================================================