+++ modules/python_engine/tests/test_type_str.m
@@ -7,11 +7,10 @@
 % SPDX-License-Identifier: LGPL-3.0-or-later
 % LICENCE_BLOCK_END
 %=============================================================================
-function r = uint16(obj)
-  if ismethod(obj, 'uint16')
-    r = obj.uint16();
-  else
-    error([_('Wrong value for #2 argument.'), ' ', 'uint16']);
-  end
-end
+% <--PYTHON ENVIRONMENT REQUIRED-->
+%=============================================================================
+R = pyrun("R = 'Hello from Python'", "R");
+assert_isequal(R.char, 'Hello from Python');
+assert_isequal(R.string, "Hello from Python");
+assert_checkerror('R.struct()', _('Wrong value for #2 argument. struct'));
 %=============================================================================