+++ CHANGELOG.md
@@ -9,9 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- [#1141](http://github.com/nelson-lang/nelson/issues/1141) Help about Managing Data between Python and Nelson.
+- Python interface (part 2):
+
+  - [#1141](http://github.com/nelson-lang/nelson/issues/1141) Help about Managing Data between Python and Nelson.
+  - [#1149](https://github.com/nelson-lang/nelson/issues/1149) python bytes, and bytearray types were not managed.
+  - `pyenv`: can use environment variables to set values.
+
 - `getenv`: Retrieve the values of several environment variables.
-- `pyenv`: can use environment variables to set values.
 
 ### Changed
 
@@ -21,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [#1144](http://github.com/nelson-lang/nelson/issues/1144) test_run markdown help file had a typo.
 - [#1143](http://github.com/nelson-lang/nelson/issues/1143) Linux Snapcraft version did not allow to use python.
+- `single(int64([1 2; 3 4]))` returned a wrong value.
+- `py.tuple`, `py.list` compatibility increased.
 
 ## 1.3.0 (2024-03-30)